requirements.txt
Django>=4.2
mysqlclient>=2.2.0
Pillow>=10.0.0
django-multiselectfield>=0.1.12
google-generativeai